refactor: 소켓 관련 파일 위치 및 이름 변경 (chat -> waiting)
