+++ src/v1/sockets/gateway.ts
@@ -1,16 +1,16 @@
 import { Server } from 'socket.io';
 import { asValue, AwilixContainer } from 'awilix';
-import startChatNamespace from './chat/startChatNamespace.js';
 import { socketMiddleware } from './utils/middleware.js';
+import startChatNamespace from './waiting/chat.namespace.js';
 
 export const registerSocketGateway = (diContainer: AwilixContainer, io: Server) => {
   io.use(socketMiddleware);
 
-  const chatNamespace = io.of('/chat');
+  const chatNamespace = io.of('/waiting');
 
   diContainer.register({
     chatNamespace: asValue(chatNamespace),
   });
 
-  startChatNamespace(io.of('/chat'));
+  startChatNamespace(io.of('/waiting'));
 };
