+++ src/v1/sockets/waiting/chat.namespace.ts
@@ -1,9 +1,8 @@
 import { Namespace, Socket } from 'socket.io';
 import * as console from 'node:console';
-import { joinRoomSchema, messageSchema } from './chat.schema.js';
 import { z } from 'zod';
-// import { socketErrorHandler } from '../utils/errorHandler.js';
 import { socketMiddleware } from '../utils/middleware.js';
+import { joinRoomSchema, messageSchema } from './waiting.schema.js';
 
 export default function startChatNamespace(namespace: Namespace) {
   namespace.use(socketMiddleware);
