+++ src/v1/sockets/waiting/waiting.schema.ts
